libcore/luni/src/test/java/org/apache/harmony/luni/tests/java/util/ArrayListTest.java
package org.apache.harmony.luni.tests.java.util;

import java.util.ArrayList;

import junit.framework.TestCase;

public class ArrayListTest extends TestCase {

    public void test_addAllCollectionOfQextendsE() {
        // Regression for HARMONY-539
        // https://issues.apache.org/jira/browse/HARMONY-539
        ArrayList<String> alist = new ArrayList<String>();
        ArrayList<String> blist = new ArrayList<String>();
        alist.add("a");
        alist.add("b");
        blist.add("c");
        blist.add("d");
        blist.remove(0);
        blist.addAll(0, alist);
        assertEquals("a", blist.get(0));
        assertEquals("b", blist.get(1));
        assertEquals("d", blist.get(2));
    }

}